move ui code in to window and calculation ish code in to calculator
