+++ BuildMonitor/LocalData/IBuildTimes.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Dynamic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -9,12 +10,12 @@ namespace BuildMonitor.LocalData
 {
     public interface IBuildTimes
     {
+        IEnumerable<ExpandoObject> SolutionMonthTable();
+
         TimeSpan Total { get; }
         IEnumerable<SolutionMonth> AvailableMonths { get; }
         IEnumerable<string> AvailableSolutions { get; }
-
         TimeSpan SolutionMonth(string solution, int month, int year);
-
         TimeSpan Solution(string solution);
     }
 }
