+++ BuildMonitorPackage/AnalyseBuildTimes.xaml
@@ -6,5 +6,5 @@
         xmlns:local="clr-namespace:BuildMonitorPackage"
         mc:Ignorable="d"
         Title="AnalyseBuildTimes" Height="300" Width="300">
-    <DataGrid x:Name="dataGrid" HorizontalAlignment="Left" Height="156" Margin="20,17,0,0" VerticalAlignment="Top" Width="266"/>
+    <DataGrid x:Name="solutionMonthDataGrid" HorizontalAlignment="Left" Height="156" Margin="20,17,0,0" VerticalAlignment="Top" Width="266"/>
 </Window>
